+++ pkgs/swift2objc/lib/src/generator/generators/class_generator.dart
@@ -13,6 +13,7 @@ import '../generator.dart';
 
 List<String> generateClass(ClassDeclaration declaration) {
   return [
+    if (declaration.isStub) ..._generateStubComment(declaration),
     ...generateAvailability(declaration),
     '${_generateClassHeader(declaration)} {',
     ...[
@@ -23,6 +24,14 @@ List<String> generateClass(ClassDeclaration declaration) {
       ..._generateNestedDeclarations(declaration),
     ].nonNulls.indent(),
     '}\n',
+  ];
+}
+
+List<String> _generateStubComment(ClassDeclaration declaration) {
+  final wrappedType = declaration.wrappedInstance!.type.swiftType;
+  return [
+    '// This wrapper is a stub. To generate the full wrapper, add $wrappedType',
+    "// to your config's include function.",
   ];
 }