+++ pkgs/swift2objc/lib/src/transformer/_core/utils.dart
@@ -18,10 +18,10 @@ import 'unique_namer.dart';
 // that weird. Refactor this as part of the transformer refactor.
 
 (String value, ReferredType type) maybeWrapValue(ReferredType type,
-    String value, UniqueNamer globalNamer, TransformationMap transformationMap,
+    String value, UniqueNamer globalNamer, TransformationState state,
     {bool shouldWrapPrimitives = false}) {
   final (wrappedPrimitiveType, returnsWrappedPrimitive) =
-      maybeGetPrimitiveWrapper(type, shouldWrapPrimitives, transformationMap);
+      maybeGetPrimitiveWrapper(type, shouldWrapPrimitives, state);
   if (returnsWrappedPrimitive) {
     return (
       '${(wrappedPrimitiveType as DeclaredType).name}($value)',
@@ -38,15 +38,14 @@ import 'unique_namer.dart';
   } else if (type is DeclaredType) {
     final declaration = type.declaration;
     if (declaration is TypealiasDeclaration) {
-      return maybeWrapValue(
-          declaration.target, value, globalNamer, transformationMap,
+      return maybeWrapValue(declaration.target, value, globalNamer, state,
           shouldWrapPrimitives: shouldWrapPrimitives);
     }
 
     final transformedTypeDeclaration = transformDeclaration(
       declaration,
       globalNamer,
-      transformationMap,
+      state,
     );
 
     return (
@@ -55,7 +54,7 @@ import 'unique_namer.dart';
     );
   } else if (type is OptionalType) {
     final (newValue, newType) =
-        maybeWrapValue(type.child, '$value!', globalNamer, transformationMap);
+        maybeWrapValue(type.child, '$value!', globalNamer, state);
     return (
       '$value == nil ? nil : $newValue',
       OptionalType(newType),