+++ pkgs/swift2objc/lib/src/transformer/transformers/transform_compound.dart
@@ -18,11 +18,9 @@ import 'transform_function.dart';
 import 'transform_initializer.dart';
 import 'transform_variable.dart';
 
-ClassDeclaration transformCompound(
-  CompoundDeclaration originalCompound,
-  UniqueNamer parentNamer,
-  TransformationMap transformationMap,
-) {
+ClassDeclaration transformCompound(CompoundDeclaration originalCompound,
+    UniqueNamer parentNamer, TransformationState state) {
+  final isStub = state.stubs.contains(originalCompound);
   final compoundNamer = UniqueNamer.inCompound(originalCompound);
 
   final wrappedCompoundInstance = PropertyDeclaration(
@@ -39,65 +37,68 @@ ClassDeclaration transformCompound(
     hasObjCAnnotation: true,
     superClass: objectType,
     isWrapper: true,
+    isStub: isStub,
     wrappedInstance: wrappedCompoundInstance,
     wrapperInitializer: buildWrapperInitializer(wrappedCompoundInstance),
   );
 
-  transformationMap[originalCompound] = transformedCompound;
+  state.map[originalCompound] = transformedCompound;
 
   transformedCompound.nestedDeclarations = originalCompound.nestedDeclarations
-      .map((nested) => maybeTransformDeclaration(
-              nested, compoundNamer, transformationMap, nested: true)
-          as InnerNestableDeclaration?)
+      .map((nested) =>
+          maybeTransformDeclaration(nested, compoundNamer, state, nested: true)
+              as InnerNestableDeclaration?)
       .nonNulls
       .toList()
     ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
   transformedCompound.nestedDeclarations
       .fillNestingParents(transformedCompound);
 
-  final transformedProperties = originalCompound.properties
-      .map((property) => transformProperty(
-            property,
-            wrappedCompoundInstance,
-            parentNamer,
-            transformationMap,
-          ))
-      .nonNulls
-      .toList();
+  if (!isStub) {
+    final transformedProperties = originalCompound.properties
+        .map((property) => transformProperty(
+              property,
+              wrappedCompoundInstance,
+              parentNamer,
+              state,
+            ))
+        .nonNulls
+        .toList();
 
-  final transformedInitializers = originalCompound.initializers
-      .map((initializer) => transformInitializer(
-            initializer,
-            wrappedCompoundInstance,
-            parentNamer,
-            transformationMap,
-          ))
-      .toList();
+    final transformedInitializers = originalCompound.initializers
+        .map((initializer) => transformInitializer(
+              initializer,
+              wrappedCompoundInstance,
+              parentNamer,
+              state,
+            ))
+        .toList();
 
-  final transformedMethods = originalCompound.methods
-      .map((method) => transformMethod(
-            method,
-            wrappedCompoundInstance,
-            parentNamer,
-            transformationMap,
-          ))
-      .nonNulls
-      .toList();
+    final transformedMethods = originalCompound.methods
+        .map((method) => transformMethod(
+              method,
+              wrappedCompoundInstance,
+              parentNamer,
+              state,
+            ))
+        .nonNulls
+        .toList();
 
-  transformedCompound.properties = transformedProperties
-      .whereType<PropertyDeclaration>()
-      .toList()
-    ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
+    transformedCompound.properties = transformedProperties
+        .whereType<PropertyDeclaration>()
+        .toList()
+      ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
 
-  transformedCompound.initializers = transformedInitializers
-      .whereType<InitializerDeclaration>()
-      .toList()
-    ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
+    transformedCompound.initializers = transformedInitializers
+        .whereType<InitializerDeclaration>()
+        .toList()
+      ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
 
-  transformedCompound.methods = (transformedMethods +
-      transformedProperties.whereType<MethodDeclaration>().toList() +
-      transformedInitializers.whereType<MethodDeclaration>().toList())
-    ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
+    transformedCompound.methods = (transformedMethods +
+        transformedProperties.whereType<MethodDeclaration>().toList() +
+        transformedInitializers.whereType<MethodDeclaration>().toList())
+      ..sort((Declaration a, Declaration b) => a.id.compareTo(b.id));
+  }
 
   return transformedCompound;
 }