+++ pkgs/swift2objc/lib/src/transformer/transformers/transform_variable.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2024, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 import '../../ast/_core/interfaces/declaration.dart';
 import '../../ast/_core/interfaces/variable_declaration.dart';
 import '../../ast/_core/shared/referred_type.dart';
@@ -20,7 +24,7 @@ Declaration? transformProperty(
   PropertyDeclaration originalProperty,
   PropertyDeclaration wrappedClassInstance,
   UniqueNamer globalNamer,
-  TransformationMap transformationMap,
+  TransformationState state,
 ) {
   if (disallowedMethods.contains(originalProperty.name)) {
     return null;
@@ -33,7 +37,7 @@ Declaration? transformProperty(
   return _transformVariable(
     originalProperty,
     globalNamer,
-    transformationMap,
+    state,
     property: true,
     wrapperPropertyName: originalProperty.name,
     variableReferenceExpression: '$propertySource.${originalProperty.name}',
@@ -43,12 +47,12 @@ Declaration? transformProperty(
 Declaration transformGlobalVariable(
   GlobalVariableDeclaration globalVariable,
   UniqueNamer globalNamer,
-  TransformationMap transformationMap,
+  TransformationState state,
 ) {
   return _transformVariable(
     globalVariable,
     globalNamer,
-    transformationMap,
+    state,
     wrapperPropertyName: globalNamer.makeUnique(
       '${globalVariable.name}Wrapper',
     ),
@@ -61,7 +65,7 @@ Declaration transformGlobalVariable(
 Declaration _transformVariable(
   VariableDeclaration originalVariable,
   UniqueNamer globalNamer,
-  TransformationMap transformationMap, {
+  TransformationState state, {
   bool property = false,
   required String wrapperPropertyName,
   required String variableReferenceExpression,
@@ -69,7 +73,7 @@ Declaration _transformVariable(
   final transformedType = transformReferredType(
     originalVariable.type,
     globalNamer,
-    transformationMap,
+    state,
   );
 
   final shouldGenerateSetter = originalVariable is PropertyDeclaration
@@ -87,7 +91,7 @@ Declaration _transformVariable(
     final resultName = localNamer.makeUnique('result');
 
     final (wrapperResult, type) = maybeWrapValue(
-        originalVariable.type, resultName, globalNamer, transformationMap,
+        originalVariable.type, resultName, globalNamer, state,
         shouldWrapPrimitives: originalVariable.throws);
 
     return MethodDeclaration(
@@ -136,7 +140,7 @@ Declaration _transformVariable(
     variableReferenceExpression,
     transformedProperty,
     globalNamer,
-    transformationMap,
+    state,
   );
   transformedProperty.getter = PropertyStatements(getterStatements);
 
@@ -146,7 +150,7 @@ Declaration _transformVariable(
       variableReferenceExpression,
       transformedProperty,
       globalNamer,
-      transformationMap,
+      state,
     );
     transformedProperty.setter = PropertyStatements(setterStatements);
   }
@@ -159,13 +163,13 @@ List<String> _generateGetterStatements(
   String variableReferenceExpression,
   PropertyDeclaration transformedProperty,
   UniqueNamer globalNamer,
-  TransformationMap transformationMap,
+  TransformationState state,
 ) {
   final (wrappedValue, wrapperType) = maybeWrapValue(
     originalVariable.type,
     variableReferenceExpression,
     globalNamer,
-    transformationMap,
+    state,
   );
 
   assert(wrapperType.sameAs(transformedProperty.type));
@@ -178,7 +182,7 @@ List<String> _generateSetterStatements(
   String variableReference,
   PropertyDeclaration transformedProperty,
   UniqueNamer globalNamer,
-  TransformationMap transformationMap,
+  TransformationState state,
 ) {
   final (unwrappedValue, unwrappedType) = maybeUnwrapValue(
     transformedProperty.type,