+++ tool/ci.dart
@@ -55,6 +55,7 @@ void main(List<String> arguments) async {
 
   if (argResults['generate'] as bool) {
     const generators = [
+      'pkgs/hooks_runner/test_data/manifest_generator.dart',
       'pkgs/hooks/tool/generate_schemas.dart',
       'pkgs/hooks/tool/generate_syntax.dart',
       'pkgs/hooks/tool/normalize.dart',
@@ -67,7 +68,11 @@ void main(List<String> arguments) async {
 
   if (argResults['test'] as bool) {
     final testUris = getTestUris(packages);
-    _runProcess('dart', ['test', ...testUris]);
+    _runProcess('dart', [
+      'test',
+      if (Platform.environment['GITHUB_ACTIONS'] != null) '--reporter=github',
+      ...testUris,
+    ]);
   }
 
   if (argResults['example'] as bool) {